+++ src/api/v1/routes/parts.ts
@@ -6,8 +6,10 @@
  * @module partsRouter
  */
 
+import { errors } from 'pg-promise';
 import { Router, Request, Response } from 'express';
 import { postgres } from '../../../db';
+import { type_guard } from '../interfaces/params';
 
 const partsRouter = Router();
 
@@ -27,18 +29,27 @@ partsRouter.get('/', async (_request:Request, response:Response):Promise<Respons
 });
 
 /** GET /v1/parts/:id */
-partsRouter.get('/get/:id', async (request:Request, response:Response):Promise<Response> => {
-    try {
-        const db = postgres.get_db();
-        const parts = await db.any(`
-            SELECT id, name, description, manufacturer_id, supplier_id, unit_price FROM parts
-            WHERE id = $[id]`,
-            { id: request.params.id }
-        );
-        response.json(parts);
-    } catch (err) {
-        console.error(err);
-        response.status(500).json({ error: err.message || err });
+partsRouter.get('/:id', async (request:Request, response:Response):Promise<Response> => {
+    if (type_guard(request.params)) {
+        try {
+            const db = postgres.get_db();
+            const parts = await db.one(`
+                SELECT id, name, description, manufacturer_id, supplier_id, unit_price FROM parts
+                WHERE id = $[id]`,
+                { id: request.params.id }
+            );
+            response.json(parts);
+        } catch (err) {
+            console.error(err);
+            // since the param is the pk more than one response row is not possible
+            if (err instanceof errors.QueryResultError) {
+                response.status(404).json({error: 'Requested supplier not found' });
+            } else {
+                response.status(500).json({ error: 'The server experienced an internal error' });
+            }
+        }
+    } else {
+        response.status(400).json({ error: 'Bad Request: id param must be an integer' });
     }
     return response;
 });
@@ -63,17 +74,21 @@ partsRouter.post('/', async (request:Request, response:Response):Promise<Respons
 
 /** DELETE /v1/parts/:id */
 partsRouter.delete('/:id', async (request:Request, response:Response):Promise<Response> => {
-    try {
-        const db = postgres.get_db();
-        const id = await db.result(`
-            DELETE FROM parts
-            WHERE id = $[id]`,
-            { id: request.params.id }, (r:any) => r.rowCount
-        );
-        response.json({ id });
-    } catch (err) {
-        console.error(err);
-        response.status(500).json({ error: err.message || err });
+    if (type_guard(request.params)) {
+        try {
+            const db = postgres.get_db();
+            const id = await db.result(`
+                DELETE FROM parts
+                WHERE id = $[id]`,
+                { id: request.params.id }, (r:any) => r.rowCount
+            );
+            response.json({ id });
+        } catch (err) {
+            console.error(err);
+            response.status(500).json({ error: err.message || err });
+        }
+    } else {
+        response.status(400).json({ error: 'Bad Request: id param must be an integer' });
     }
     return response;
 });
