+++ src/api/v1/routes/projects.ts
@@ -6,6 +6,7 @@
  * @module projectsRouter
  */
 
+import { errors } from 'pg-promise';
 import { Router, Request, Response } from 'express';
 import { postgres } from '../../../db';
 import { type_guard } from '../interfaces/params';
@@ -30,17 +31,26 @@ projectsRouter.get('/', async (_request:Request, response:Response):Promise<Resp
 
 /** GET /v1/projects/:id */
 projectsRouter.get('/:id', async (request:Request, response:Response):Promise<Response> => {
-    try {
-        const db = postgres.get_db();
-        const projects = await db.any(`
-            SELECT id, name, description FROM projects
-            WHERE id = $[id]`,
-            { id: request.params.id }
-        );
-        response.json(projects);
-    } catch (err) {
-        console.error(err);
-        response.status(500).json({ error: err.message || err });
+    if (type_guard(request.params)) {
+        try {
+            const db = postgres.get_db();
+            const projects = await db.one(`
+                SELECT id, name, description FROM projects
+                WHERE id = $[id]`,
+                { id: request.params.id }
+            );
+            response.json(projects);
+        } catch (err) {
+            console.error(err);
+            // since the param is the pk more than one response row is not possible
+            if (err instanceof errors.QueryResultError) {
+                response.status(404).json({error: 'Project not found' });
+            } else {
+                response.status(500).json({ error: 'The server experienced an internal error' });
+            }
+        }
+    } else {
+        response.status(400).json({ error: 'Bad Request: id param must be an integer' });
     }
     return response;
 });
@@ -65,17 +75,21 @@ projectsRouter.post('/', async (request:Request, response:Response):Promise<Resp
 
 /** DELETE /v1/projects/:id */
 projectsRouter.delete('/:id', async (request:Request, response:Response):Promise<Response> => {
-    try {
-        const db = postgres.get_db();
-        const id = await db.result(`
-            DELETE FROM projects
-            WHERE id = $[id]`,
-            { id: request.params.id }, (r:any) => r.rowCount
-        );
-        response.json({ id });
-    } catch (err) {
-        console.error(err);
-        response.status(500).json({ error: err.message || err });
+    if (type_guard(request.params)) {
+        try {
+            const db = postgres.get_db();
+            const id = await db.result(`
+                DELETE FROM projects
+                WHERE id = $[id]`,
+                { id: request.params.id }, (r:any) => r.rowCount
+            );
+            response.json({ id });
+        } catch (err) {
+            console.error(err);
+            response.status(500).json({ error: err.message || err });
+        }
+    } else {
+        response.status(400).json({ error: 'Bad Request: id param must be an integer' });
     }
     return response;
 });
